event_crawler/backend/requirements.txt
Flask==2.3.2
flask_cors==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6